Introduced Response class and integrated it into MainException and Controllers
A new Response class has been added to extend response capability, including features to output JSON data and render file content. This class has been integrated into the MainException and Controller classes to improve response handling. The kernel was also refactored to respond with the specified file content when the controller class does not exist instead of logging a warning and throwing a NotFoundException